scripts/requirements.txt
lxml
pyyaml
